Add API to re-order columns in pgdump schema SQL files
We currently collapse migrations to a single migration file, which means
that when we bring up a new shard we generally properly trigger the
collapsing logic since 1.) it goes through the ActiveRecord migration
code we've overridden and 2.) we get most of the possible benefits since
by collapsing we've moved the accretion of columns over time into a
single `create_table` call.

But as we move towards collapsing to SQL files instead, we no longer
have an intercept point in the ActiveRecord migration codepath. Instead
we need to take the SQL schema files we've dumped via the pgdump utility
and re-order the columns in that SQL file to be byte-packed.