More float ordering fixes
While applying this to the gateway, I noticed that 8-byte floats
appeared to be out of expected order. There's a bit of a rabbit-hole
here: ActiveRecord allows supplying a `limit` option for floats (which
in part allows you to implicitly control the width of the column).
Unlike integers, however, where the the limit is the number of bytes,
the limit on floats is the number of bits of precisision (in other words
you're controlling the split in storage between the exponent and
mantissa).

Anyway, this change should handle both of those cases properly.