Handle quoted column names
Column names in a `CREATE TABLE` statement may be quoted (if the name is
a reserved word or if it contains spaces or special characters). While
the reserved word case by itself would be "easy" to handle with some
more regex'ing, this seems like it's getting to the point where you're
clearly implementing a parser. So, absent a small parser that just
handles quoted string parsing, I decided to use the PostgreSQL parsing
code itself (by way of the pg_query gem) -- but on a per-column basis in
a bit of a contrived fashion so as to avoid needing to go back from
parse tree to updated DDL statement.