Fix double precision/real handling in pgdump output
When I'd initially copied the tests over for the pgdump support I'd
thought we were handling this case, but forgot that in ActiveRecord
"decimal" is a floating point value while in PostgreSQL "decimal" is
"numeric", so the tests for "decimal" weren't sufficient. Subsequently
I'd added tests to the ActiveRecord side to tests floats with limits
specifically, but didn't realize we were still missing support on the
pgdump side.

Here we add that support.