Support type modifiers generically
We were already handling types like `varchar(n)`, but for types where we
didn't have something special to do with the modifier, we were trying to
query `pg_type` for the name with the modifier, which obviously won't
work. Instead we now look up the canonical name without regard for the
modifier.

I also shored up the tests for the fact that ActiveRecord canonicalizes
both decimal and numeric to decimal, while Postgres does exactly the
opposite.