Properly support bit/character and bit(n)/character(n) types
The code looked like it already handled these types, but there are a
couple of issues:

First, as we already knew (this is clear in relevant comments in the
code) there's a mismatch between how the type name is canonicalized in
Postgres for external consumption and the pg_type typname entry. For
example, pg_dump (and `pg_catalog.format_type()`) outputs character
types as `character`, but the relevant `typname` entry in `pg_type` is
`char`. This is similarly true with `varchar`, but we already handle
those by name explicitly, so it didn't matter for our purposes, but
`varbit` has the same concerns.

Second, due to the vagaries of the SQL spec, the double-quoted variants
of `"bit"` and `"char"` are _not_ equivalent to the unquoted variants
(specifically they match up to single-length types).