Support time[tz] data type
I believe this is the last of the built-in data types that have special
representations in their formatted form that we don't already support.
Just like timestamp the `with time zone` and `without time zone`
qualifiers are expected in the formatted form.

Along the way I made the timestamp type name matcher a bit more specific
to avoid accidentally matching user types.