Fix ordering of booleans
There was a long-standing bug here which treated booleans and smallints
the same way, but booleans only take up 1 byte, not two, and they don't
need any alignment, so treating them with the 2 byte stored (and
aligned) smallint isn't correct. It's possible this is a holdover from
way-back-when and it was originally just a rough ordering not directly
based on actual alignment, but in any case, the comment about it being
the actual required alignment wasn't correct.